refactor(sensor_board): :recycle: move logging out of main.c
